requirements.txt
argon2==0.1.10
argon2-cffi==19.2.0
asgiref==3.2.7
beautifulsoup4==4.9.0
cffi==1.14.0
dj-database-url==0.5.0
Django==3.0.6
django-bootstrap4==1.1.1
django-braces==1.14.0
django-filter==2.2.0
django-widget-tweaks==1.4.8
Faker==4.0.3
gunicorn==20.0.4
misaka==2.1.1
Pillow==7.1.2
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.2.7
pytz==2020.1
sendgrid==6.3.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.0.1
